+++ test/specs/login.e2e.ts
@@ -24,5 +24,3 @@ describe('Login with correct username and password', () => {
         );
     });
 });
-
-
